# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 2Grils.PPT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates with filtering and sorting
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: slideCount
          in: query
          schema:
            type: integer
        - name: aspectRatio
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, best-seller, price-asc, price-desc, top-rated]
        - name: search
          in: query
          schema:
            type: string
        - name: isFree
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
...
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /templates/{id}/related:
    get:
      operationId: getRelatedTemplates
      tags: [templates]
      summary: Get related templates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
  /templates/{id}/reviews:
    get:
      operationId: listTemplateReviews
      tags: [templates]
      summary: Get reviews for a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [templates]
      summary: Submit a review for a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Categories
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/payment-status:
    get:
      operationId: getOrderPaymentStatus
      tags: [orders]
      summary: Poll payment status for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"

  # User library
...
  /wishlist/{templateId}:
    post:
      operationId: addToWishlist
      tags: [templates]
      summary: Add template to wishlist
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFromWishlist
      tags: [templates]
      summary: Remove template from wishlist
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Custom Design Requests
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Template:
      type: object
      required: [id, titleVi, titleEn, slug, price, thumbnailUrl, slideCount, aspectRatio, categoryId, categoryName, style, isFeatured, isBestSeller, avgRating, reviewCount, salesCount, createdAt]
      properties:
        id:
          type: integer
        titleVi:
          type: string
        titleEn:
          type: string
        slug:
          type: string
        price:
          type: number
        isFree:
          type: boolean
        thumbnailUrl:
          type: string
        previewImages:
          type: array
          items:
            type: string
        slideCount:
          type: integer
        aspectRatio:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        style:
          type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isBestSeller:
          type: boolean
        avgRating:
          type: number
        reviewCount:
          type: integer
        salesCount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
    TemplateDetail:
      allOf:
        - $ref: "#/components/schemas/Template"
        - type: object
          properties:
            descriptionVi:
              type: string
            descriptionEn:
              type: string
            features:
              type: array
              items:
                type: string
            compatibleSoftware:
              type: array
              items:
                type: string
            ratingDistribution:
              type: array
              items:
                $ref: "#/components/schemas/RatingDistribution"
    RatingDistribution:
      type: object
      required: [stars, count]
      properties:
        stars:
          type: integer
        count:
          type: integer
    TemplateListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Template"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Review:
      type: object
      required: [id, templateId, authorName, rating, comment, isVerifiedPurchase, createdAt, isHidden]
      properties:
        id:
          type: integer
        templateId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        imageUrl:
          type: ["string", "null"]
        isVerifiedPurchase:
          type: boolean
        isHidden:
          type: boolean
        createdAt:
          type: string
    ReviewListResponse:
      type: object
      required: [items, avgRating, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        avgRating:
          type: number
        total:
          type: integer
        ratingDistribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingDistribution"
    CreateReviewBody:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
        comment:
          type: string
        imageUrl:
          type: string
...
    CartItem:
      type: object
      required: [templateId, titleVi, titleEn, price, thumbnailUrl]
      properties:
        templateId:
          type: integer
        titleVi:
          type: string
        titleEn:
          type: string
        price:
          type: number
        thumbnailUrl:
          type: string
...
    Order:
      type: object
      required: [id, status, total, currency, items, customerName, customerEmail, createdAt, paymentMethod]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, paid, failed, refunded]
        total:
          type: number
        currency:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: ["string", "null"]
        paymentMethod:
          type: string
        qrCode:
          type: ["string", "null"]
        transferContent:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
        downloadLinks:
          type: array
          items:
            type: string
    CreateOrderBody:
      type: object
      required: [customerName, customerEmail]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        voucherCode:
          type: string
        needVatInvoice:
          type: boolean
        companyName:
          type: string
        taxCode:
          type: string
    PaymentStatus:
      type: object
      required: [orderId, status]
      properties:
        orderId:
          type: integer
        status:
          type: string
          enum: [pending, paid, failed, expired]
        paidAt:
          type: ["string", "null"]
...